improvement: add TapbackResponses view
