+++ Meshtastic/Views/Messages/ChannelMessageList.swift
@@ -61,42 +61,10 @@ struct ChannelMessageList: View {
 										self.messageFieldFocused = true
 									}
 
-									let tapbacks = message.value(forKey: "tapbacks") as? [MessageEntity] ?? []
-									if tapbacks.count > 0 {
-										VStack(alignment: .trailing) {
-											HStack {
-												ForEach( tapbacks ) { (tapback: MessageEntity) in
-													VStack {
-														let image = tapback.messagePayload!.image(fontSize: 20)
-														Image(uiImage: image!).font(.caption)
-														Text("\(tapback.fromUser?.shortName ?? "?")")
-															.font(.caption2)
-															.foregroundColor(.gray)
-															.fixedSize()
-															.padding(.bottom, 1)
-													}
-													.onAppear {
-														if !tapback.read {
-															tapback.read = true
-															do {
-																try context.save()
-																print("📖 Read message \(message.messageId) ")
-																appState.unreadChannelMessages = myInfo.unreadMessages
-																UIApplication.shared.applicationIconBadgeNumber = appState.unreadChannelMessages + appState.unreadDirectMessages
-																context.refresh(myInfo, mergeChanges: true)
-															} catch {
-																print("Failed to read tapback \(tapback.messageId)")
-															}
-														}
-													}
-												}
-											}
-											.padding(10)
-											.overlay(
-												RoundedRectangle(cornerRadius: 18)
-													.stroke(Color.gray, lineWidth: 1)
-											)
-										}
+									TapbackResponses(message: message) {
+										appState.unreadChannelMessages = myInfo.unreadMessages
+										UIApplication.shared.applicationIconBadgeNumber = appState.unreadChannelMessages + appState.unreadDirectMessages
+										context.refresh(myInfo, mergeChanges: true)
 									}
 					
 									HStack {
